[Refactor] Move DataSegment out of chunks sorter
