+++ be/src/compute_env/sorting/data_segment.h
@@ -1,0 +1,70 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <cstddef>
+#include <cstdint>
+#include <memory>
+#include <vector>
+
+#include "column/chunk.h"
+#include "column/vectorized_fwd.h"
+
+namespace starrocks {
+
+class ExprContext;
+
+struct DataSegment {
+    static const uint8_t SMALLER_THAN_MIN_OF_SEGMENT = 2;
+    static const uint8_t INCLUDE_IN_SEGMENT = 1;
+    static const uint8_t LARGER_THAN_MAX_OF_SEGMENT = 0;
+
+    ChunkPtr chunk;
+    Columns order_by_columns;
+
+    DataSegment() : chunk(std::make_shared<Chunk>()) {}
+
+    DataSegment(const std::vector<ExprContext*>* sort_exprs, const ChunkPtr& cnk) { init(sort_exprs, cnk); }
+
+    int64_t mem_usage() const { return chunk->memory_usage(); }
+
+    void init(const std::vector<ExprContext*>* sort_exprs, const ChunkPtr& cnk);
+
+    void clear() {
+        chunk.reset(std::make_unique<Chunk>().release());
+        order_by_columns.clear();
+    }
+
+    // Return value:
+    //  < 0: current row precedes the row in the other chunk;
+    // == 0: current row is equal to the row in the other chunk;
+    //  > 0: current row succeeds the row in the other chunk;
+    int compare_at(size_t index_in_chunk, const DataSegment& other, size_t index_in_other_chunk,
+                   const std::vector<int>& sort_order_flag, const std::vector<int>& null_first_flag) const {
+        size_t col_number = order_by_columns.size();
+        for (size_t col_index = 0; col_index < col_number; ++col_index) {
+            const auto& left_col = order_by_columns[col_index];
+            const auto& right_col = other.order_by_columns[col_index];
+            int c = left_col->compare_at(index_in_chunk, index_in_other_chunk, *right_col, null_first_flag[col_index]);
+            if (c != 0) {
+                return c * sort_order_flag[col_index];
+            }
+        }
+        return 0;
+    }
+};
+using DataSegments = std::vector<DataSegment>;
+
+} // namespace starrocks
