+++ be/src/exec/chunks_sorter_full_sort.cpp
@@ -17,6 +17,7 @@
 #include "base/failpoint/fail_point.h"
 #include "base/utility/defer_op.h"
 #include "common/runtime_profile.h"
+#include "compute_env/sorting/data_segment.h"
 #include "compute_env/sorting/merge.h"
 #include "compute_env/sorting/sort_permute.h"
 #include "compute_env/sorting/sorting.h"
