+++ be/src/exec/chunks_sorter_heap_sort.h
@@ -23,6 +23,7 @@
 #include "column/chunk.h"
 #include "column/vectorized_fwd.h"
 #include "common/object_pool.h"
+#include "compute_env/sorting/data_segment.h"
 #include "exec/chunks_sorter.h"
 #include "exprs/expr_context.h"
 #include "glog/logging.h"
