+++ be/src/exec/chunks_sorter_topn.h
@@ -16,6 +16,7 @@
 
 #include "column/vectorized_fwd.h"
 #include "common/runtime_profile.h"
+#include "compute_env/sorting/data_segment.h"
 #include "compute_env/sorting/merge.h"
 #include "exec/chunks_sorter.h"
 #include "exprs/expr_context.h"
