+++ be/src/exec/spill/mem_table.cpp
@@ -21,10 +21,11 @@
 #include "column/chunk.h"
 #include "column/vectorized_fwd.h"
 #include "common/runtime_profile.h"
+#include "compute_env/sorting/data_segment.h"
 #include "compute_env/workgroup/scan_task.h"
-#include "exec/chunks_sorter.h"
 #include "exec/spill/input_stream.h"
 #include "exec/spill/serde.h"
+#include "exec/spill/spiller.h"
 #include "exec/spill/task_executor.h"
 #include "exec/spill/yield.h"
 #include "runtime/current_thread.h"
